build base HAL service
