+++ itlwm/hal_iwm/phy.cpp
@@ -119,9 +119,9 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-#include "itlwm.hpp"
+#include "ItlIwm.hpp"
 
-struct iwm_phy_db_entry * itlwm::
+struct iwm_phy_db_entry * ItlIwm::
 iwm_phy_db_get_section(struct iwm_softc *sc, uint16_t type, uint16_t chg_id)
 {
     struct iwm_phy_db *phy_db = &sc->sc_phy_db;
@@ -148,7 +148,7 @@ iwm_phy_db_get_section(struct iwm_softc *sc, uint16_t type, uint16_t chg_id)
     return NULL;
 }
 
-int itlwm::
+int ItlIwm::
 iwm_phy_db_set_section(struct iwm_softc *sc,
                        struct iwm_calib_res_notif_phy_db *phy_db_notif)
 {
@@ -178,7 +178,7 @@ iwm_phy_db_set_section(struct iwm_softc *sc,
     return 0;
 }
 
-int itlwm::
+int ItlIwm::
 iwm_phy_db_get_section_data(struct iwm_softc *sc, uint32_t type, uint8_t **data,
                             uint16_t *size, uint16_t ch_id)
 {
@@ -200,7 +200,7 @@ iwm_phy_db_get_section_data(struct iwm_softc *sc, uint32_t type, uint8_t **data,
     return 0;
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_phy_db_cmd(struct iwm_softc *sc, uint16_t type, uint16_t length,
                     void *data)
 {
@@ -222,7 +222,7 @@ iwm_send_phy_db_cmd(struct iwm_softc *sc, uint16_t type, uint16_t length,
     return iwm_send_cmd(sc, &cmd);
 }
 
-int itlwm::
+int ItlIwm::
 iwm_phy_db_send_all_channel_groups(struct iwm_softc *sc, uint16_t type,
                                    uint8_t max_ch_groups)
 {
@@ -249,7 +249,7 @@ iwm_phy_db_send_all_channel_groups(struct iwm_softc *sc, uint16_t type,
     return 0;
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_phy_db_data(struct iwm_softc *sc)
 {
     XYLog("%s\n", __FUNCTION__);
@@ -294,7 +294,7 @@ iwm_send_phy_db_data(struct iwm_softc *sc)
 #define IWM_ROC_TE_TYPE_NORMAL IWM_TE_P2P_DEVICE_DISCOVERABLE
 #define IWM_ROC_TE_TYPE_MGMT_TX IWM_TE_P2P_CLIENT_ASSOC
 
-int itlwm::
+int ItlIwm::
 iwm_send_time_event_cmd(struct iwm_softc *sc,
                         const struct iwm_time_event_cmd *cmd)
 {
@@ -336,7 +336,7 @@ out:
     return err;
 }
 
-void itlwm::
+void ItlIwm::
 iwm_protect_session(struct iwm_softc *sc, struct iwm_node *in,
                     uint32_t duration, uint32_t max_delay)
 {
@@ -372,7 +372,7 @@ iwm_protect_session(struct iwm_softc *sc, struct iwm_node *in,
     DELAY(100);
 }
 
-void itlwm::
+void ItlIwm::
 iwm_unprotect_session(struct iwm_softc *sc, struct iwm_node *in)
 {
     struct iwm_time_event_cmd time_cmd;
@@ -394,7 +394,7 @@ iwm_unprotect_session(struct iwm_softc *sc, struct iwm_node *in)
     DELAY(100);
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_phy_cfg_cmd(struct iwm_softc *sc)
 {
     XYLog("%s\n", __FUNCTION__);
@@ -411,7 +411,7 @@ iwm_send_phy_cfg_cmd(struct iwm_softc *sc)
                             sizeof(phy_cfg_cmd), &phy_cfg_cmd);
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_dqa_cmd(struct iwm_softc *sc)
 {
     struct iwm_dqa_enable_cmd dqa_cmd = {
@@ -423,7 +423,7 @@ iwm_send_dqa_cmd(struct iwm_softc *sc)
     return iwm_send_cmd_pdu(sc, cmd_id, 0, sizeof(dqa_cmd), &dqa_cmd);
 }
 
-int itlwm::
+int ItlIwm::
 iwm_binding_cmd(struct iwm_softc *sc, struct iwm_node *in, uint32_t action)
 {
     struct iwm_binding_cmd cmd;
@@ -460,7 +460,7 @@ iwm_binding_cmd(struct iwm_softc *sc, struct iwm_node *in, uint32_t action)
     return err;
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_cmd(struct iwm_softc *sc, struct iwm_host_cmd *hcmd)
 {
     struct iwm_tx_ring *ring = &sc->txq[sc->cmdqid];
@@ -639,7 +639,7 @@ out:
     return err;
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_cmd_pdu(struct iwm_softc *sc, uint32_t id, uint32_t flags,
                  uint16_t len, const void *data)
 {
@@ -653,7 +653,7 @@ iwm_send_cmd_pdu(struct iwm_softc *sc, uint32_t id, uint32_t flags,
     return iwm_send_cmd(sc, &cmd);
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_cmd_status(struct iwm_softc *sc, struct iwm_host_cmd *cmd,
                     uint32_t *status)
 {
@@ -685,7 +685,7 @@ iwm_send_cmd_status(struct iwm_softc *sc, struct iwm_host_cmd *cmd,
     return err;
 }
 
-int itlwm::
+int ItlIwm::
 iwm_send_cmd_pdu_status(struct iwm_softc *sc, uint32_t id, uint16_t len,
                         const void *data, uint32_t *status)
 {
@@ -698,7 +698,7 @@ iwm_send_cmd_pdu_status(struct iwm_softc *sc, uint32_t id, uint16_t len,
     return iwm_send_cmd_status(sc, &cmd, status);
 }
 
-void itlwm::
+void ItlIwm::
 iwm_free_resp(struct iwm_softc *sc, struct iwm_host_cmd *hcmd)
 {
     _KASSERT((hcmd->flags & (IWM_CMD_WANT_RESP)) == IWM_CMD_WANT_RESP);
@@ -706,7 +706,7 @@ iwm_free_resp(struct iwm_softc *sc, struct iwm_host_cmd *hcmd)
     hcmd->resp_pkt = NULL;
 }
 
-void itlwm::
+void ItlIwm::
 iwm_cmd_done(struct iwm_softc *sc, int qid, int idx, int code)
 {
     struct iwm_tx_ring *ring = &sc->txq[sc->cmdqid];
