+++ itlwm/itlwm.cpp
@@ -74,17 +74,17 @@ IOService* itlwm::probe(IOService *provider, SInt32 *score)
     if (!device) {
         return NULL;
     }
-    if (iwm_match(device)) {
-        device->findPCICapability(PCI_CAP_ID_MSI, &msiCap);
-        if (msiCap) {
-            pciMsiSetEnable(device, msiCap, 0);
-        }
-        device->findPCICapability(PCI_CAP_ID_MSIX, &msixCap);
-        if (msixCap) {
-            pciMsiXClearAndSet(device, msixCap, PCI_MSIX_FLAGS_ENABLE, 0);
-        }
-        return this;
-    }
+//    if (iwm_match(device)) {
+//        device->findPCICapability(PCI_CAP_ID_MSI, &msiCap);
+//        if (msiCap) {
+//            pciMsiSetEnable(device, msiCap, 0);
+//        }
+//        device->findPCICapability(PCI_CAP_ID_MSIX, &msixCap);
+//        if (msixCap) {
+//            pciMsiXClearAndSet(device, msixCap, PCI_MSIX_FLAGS_ENABLE, 0);
+//        }
+//        return this;
+//    }
     return NULL;
 }
 
@@ -101,8 +101,8 @@ bool itlwm::configureInterface(IONetworkInterface *netif) {
         XYLog("network statistics buffer unavailable?\n");
         return false;
     }
-    com.sc_ic.ic_ac.ac_if.netStat = fpNetStats;
-    com.sc_ic.ic_ac.ac_if.iface = OSDynamicCast(IOEthernetInterface, netif);
+    fHalService->get80211Controller()->ic_ac.ac_if.netStat = fpNetStats;
+    fHalService->get80211Controller()->ic_ac.ac_if.iface = OSDynamicCast(IOEthernetInterface, netif);
     fpNetStats->collisions = 0;
     
     return true;
@@ -123,12 +123,12 @@ IONetworkInterface *itlwm::createInterface()
 
 struct ifnet *itlwm::getIfp()
 {
-    return &com.sc_ic.ic_ac.ac_if;
+    return &fHalService->get80211Controller()->ic_ac.ac_if;
 }
 
 struct iwm_softc *itlwm::getSoft()
 {
-    return &com;
+    return NULL;
 }
 
 IOEthernetInterface *itlwm::getNetworkInterface()
@@ -169,7 +169,7 @@ ieee80211_join join;
 
 void itlwm::joinSSID(const char *ssid_name, const char *ssid_pwd)
 {
-    struct ieee80211com *ic = &com.sc_ic;
+    struct ieee80211com *ic = fHalService->get80211Controller();
     
     if (strlen(ssid_pwd) == 0) {
         memset(&nwkey, 0, sizeof(ieee80211_nwkey));
@@ -210,7 +210,7 @@ struct ieee80211_nwid nwid;
 
 void itlwm::associateSSID(const char *ssid, const char *pwd)
 {
-    struct ieee80211com *ic = &com.sc_ic;
+    struct ieee80211com *ic = fHalService->get80211Controller();
     if (strlen(pwd) == 0) {
         memcpy(nwid.i_nwid, ssid, 32);
         nwid.i_len = strlen((char *)nwid.i_nwid);
@@ -278,7 +278,8 @@ void itlwm::associateSSID(const char *ssid, const char *pwd)
         ieee80211_new_state(ic, IEEE80211_S_SCAN, -1);
     } else {
         ieee80211_node_join_bss(ic, selbs, 1);
-        com.sc_flags &= ~(IWM_FLAG_SCANNING | IWM_FLAG_BGSCAN);
+        /* TODO */
+//        com.sc_flags &= ~(IWM_FLAG_SCANNING | IWM_FLAG_BGSCAN);
     }
 }
 
@@ -308,7 +309,7 @@ bool itlwm::start(IOService *provider)
         releaseAll();
         return false;
     }
-    _fCommandGate = IOCommandGate::commandGate(this, (IOCommandGate::Action)tsleepHandler);
+    _fCommandGate = IOCommandGate::commandGate(this, (IOCommandGate::Action)itlwm::tsleepHandler);
     if (_fCommandGate == 0) {
         XYLog("No command gate!!\n");
         releaseAll();
@@ -321,9 +322,7 @@ bool itlwm::start(IOService *provider)
         releaseAll();
         return false;
     }
-    pci.workloop = _fWorkloop;
-    pci.pa_tag = pciNub;
-    if (!iwm_attach(&com, &pci)) {
+    if (!fHalService->attach(pciNub)) {
         releaseAll();
         return false;
     }
@@ -337,7 +336,7 @@ bool itlwm::start(IOService *provider)
         releaseAll();
         return false;
     }
-    watchdogTimer = IOTimerEventSource::timerEventSource(this, OSMemberFunctionCast(IOTimerEventSource::Action, this, &itlwm::watchdogAction));
+    watchdogTimer = IOTimerEventSource::timerEventSource(this, OSMemberFunctionCast(IOTimerEventSource::Action, fHalService, &ItlHalService::watchdogAction));
     if (!watchdogTimer) {
         XYLog("init watchdog fail\n");
         releaseAll();
@@ -370,18 +369,12 @@ bool itlwm::start(IOService *provider)
         }
         iterator->release();
     }
-    if (TAILQ_EMPTY(&com.sc_ic.ic_ess)) {
-        com.sc_ic.ic_flags |= IEEE80211_F_AUTO_JOIN;
+    if (TAILQ_EMPTY(&fHalService->get80211Controller()->ic_ess)) {
+        fHalService->get80211Controller()->ic_flags |= IEEE80211_F_AUTO_JOIN;
     }
     registerService();
     fNetIf->registerService();
     return true;
-}
-
-void itlwm::watchdogAction(IOTimerEventSource *timer)
-{
-    iwm_watchdog(&com.sc_ic.ic_ac.ac_if);
-    watchdogTimer->setTimeoutMS(1000);
 }
 
 const OSString * itlwm::newVendorString() const
@@ -439,7 +432,7 @@ IOReturn itlwm::selectMedium(const IONetworkMedium *medium) {
 void itlwm::stop(IOService *provider)
 {
     XYLog("%s\n", __FUNCTION__);
-    struct ifnet *ifp = &com.sc_ic.ic_ac.ac_if;
+    struct ifnet *ifp = &fHalService->get80211Controller()->ic_ac.ac_if;
     
     super::stop(provider);
     setLinkStatus(kIONetworkLinkValid);
@@ -447,58 +440,58 @@ void itlwm::stop(IOService *provider)
     detachInterface(fNetIf);
     OSSafeReleaseNULL(fNetIf);
     ifp->iface = NULL;
-    taskq_destroy(systq);
-    taskq_destroy(com.sc_nswq);
+//    taskq_destroy(systq);
+//    taskq_destroy(com.sc_nswq);
     releaseAll();
 }
 
 void itlwm::releaseAll()
 {
-    pci_intr_handle *intrHandler = com.ih;
-    
-    if (com.sc_calib_to) {
-        timeout_del(&com.sc_calib_to);
-        timeout_free(&com.sc_calib_to);
-    }
-    if (com.sc_led_blink_to) {
-        timeout_del(&com.sc_led_blink_to);
-        timeout_free(&com.sc_led_blink_to);
-    }
-    if (_fWorkloop) {
-        if (intrHandler) {
-            if (intrHandler->intr) {
-                intrHandler->intr->disable();
-                intrHandler->workloop->removeEventSource(intrHandler->intr);
-                intrHandler->intr->release();
-            }
-            intrHandler->intr = NULL;
-            intrHandler->workloop = NULL;
-            intrHandler->arg = NULL;
-            intrHandler->dev = NULL;
-            intrHandler->func = NULL;
-            intrHandler->release();
-            intrHandler = NULL;
-        }
-        if (_fCommandGate) {
-            if (lastSleepChan) {
-                wakeupOn(lastSleepChan);
-            }
-//            _fCommandGate->disable();
-            _fWorkloop->removeEventSource(_fCommandGate);
-            _fCommandGate->release();
-            _fCommandGate = NULL;
-        }
-        if (fWatchdogWorkLoop && watchdogTimer) {
-            watchdogTimer->cancelTimeout();
-            fWatchdogWorkLoop->removeEventSource(watchdogTimer);
-            watchdogTimer->release();
-            watchdogTimer = NULL;
-            fWatchdogWorkLoop->release();
-            fWatchdogWorkLoop = NULL;
-        }
-        _fWorkloop->release();
-        _fWorkloop = NULL;
-    }
+//    pci_intr_handle *intrHandler = com.ih;
+//
+//    if (com.sc_calib_to) {
+//        timeout_del(&com.sc_calib_to);
+//        timeout_free(&com.sc_calib_to);
+//    }
+//    if (com.sc_led_blink_to) {
+//        timeout_del(&com.sc_led_blink_to);
+//        timeout_free(&com.sc_led_blink_to);
+//    }
+//    if (_fWorkloop) {
+//        if (intrHandler) {
+//            if (intrHandler->intr) {
+//                intrHandler->intr->disable();
+//                intrHandler->workloop->removeEventSource(intrHandler->intr);
+//                intrHandler->intr->release();
+//            }
+//            intrHandler->intr = NULL;
+//            intrHandler->workloop = NULL;
+//            intrHandler->arg = NULL;
+//            intrHandler->dev = NULL;
+//            intrHandler->func = NULL;
+//            intrHandler->release();
+//            intrHandler = NULL;
+//        }
+//        if (_fCommandGate) {
+//            if (lastSleepChan) {
+//                wakeupOn(lastSleepChan);
+//            }
+////            _fCommandGate->disable();
+//            _fWorkloop->removeEventSource(_fCommandGate);
+//            _fCommandGate->release();
+//            _fCommandGate = NULL;
+//        }
+//        if (fWatchdogWorkLoop && watchdogTimer) {
+//            watchdogTimer->cancelTimeout();
+//            fWatchdogWorkLoop->removeEventSource(watchdogTimer);
+//            watchdogTimer->release();
+//            watchdogTimer = NULL;
+//            fWatchdogWorkLoop->release();
+//            fWatchdogWorkLoop = NULL;
+//        }
+//        _fWorkloop->release();
+//        _fWorkloop = NULL;
+//    }
     unregistPM();
 }
 
@@ -515,11 +508,11 @@ void itlwm::free()
 IOReturn itlwm::enable(IONetworkInterface *netif)
 {
     XYLog("%s\n", __FUNCTION__);
-    struct ifnet *ifp = &com.sc_ic.ic_ac.ac_if;
+    struct ifnet *ifp = &fHalService->get80211Controller()->ic_ac.ac_if;
     super::enable(netif);
     ifp->if_flags |= IFF_UP;
     _fCommandGate->enable();
-    iwm_activate(&com, DVACT_WAKEUP);
+    fHalService->enable(netif);
     watchdogTimer->setTimeoutMS(1000);
     watchdogTimer->enable();
     return kIOReturnSuccess;
@@ -531,16 +524,16 @@ IOReturn itlwm::disable(IONetworkInterface *netif)
     super::disable(netif);
     watchdogTimer->cancelTimeout();
     watchdogTimer->disable();
-    iwm_activate(&com, DVACT_QUIESCE);
+    fHalService->disable(netif);
     setLinkStatus(kIONetworkLinkValid);
     return kIOReturnSuccess;
 }
 
 IOReturn itlwm::getHardwareAddress(IOEthernetAddress *addrP) {
-    if (IEEE80211_ADDR_EQ(etheranyaddr, com.sc_ic.ic_myaddr)) {
+    if (IEEE80211_ADDR_EQ(etheranyaddr, fHalService->get80211Controller()->ic_myaddr)) {
         return kIOReturnError;
     } else {
-        IEEE80211_ADDR_COPY(addrP, com.sc_ic.ic_myaddr);
+        IEEE80211_ADDR_COPY(addrP, fHalService->get80211Controller()->ic_myaddr);
         return kIOReturnSuccess;
     }
 }
@@ -548,9 +541,9 @@ IOReturn itlwm::getHardwareAddress(IOEthernetAddress *addrP) {
 UInt32 itlwm::outputPacket(mbuf_t m, void *param)
 {
 //    XYLog("%s\n", __FUNCTION__);
-    ifnet *ifp = &com.sc_ic.ic_ac.ac_if;
+    ifnet *ifp = &fHalService->get80211Controller()->ic_ac.ac_if;
     
-    if (com.sc_ic.ic_state != IEEE80211_S_RUN || ifp == NULL || ifp->if_snd == NULL) {
+    if (fHalService->get80211Controller()->ic_state != IEEE80211_S_RUN || ifp == NULL || ifp->if_snd == NULL) {
         freePacket(m);
         return kIOReturnOutputDropped;
     }
@@ -608,38 +601,14 @@ IOReturn itlwm::getPacketFilters(const OSSymbol *group, UInt32 *filters) const {
     return rtn;
 }
 
-void itlwm::wakeupOn(void *ident)
-{
-//    XYLog("%s\n", __FUNCTION__);
-    if (_fCommandGate == 0)
-        return;
-    else
-        _fCommandGate->commandWakeup(ident);
+IOReturn itlwm::getMaxPacketSize(UInt32 *maxSize) const {
+    return super::getMaxPacketSize(maxSize);
 }
 
-int itlwm::tsleep_nsec(void *ident, int priority, const char *wmesg, int timo)
+IOReturn itlwm::
+tsleepHandler(OSObject* owner, void* arg0, void* arg1, void* arg2, void* arg3)
 {
-//    XYLog("%s %s\n", __FUNCTION__, wmesg);
-    IOReturn ret;
-    if (_fCommandGate == 0) {
-        IOSleep(timo);
-        return 0;
-    }
-    lastSleepChan = ident;
-    if (timo == 0) {
-        ret = _fCommandGate->runCommand(ident);
-    } else {
-        ret = _fCommandGate->runCommand(ident, &timo);
-    }
-    if (ret == kIOReturnSuccess)
-        return 0;
-    else
-        return 1;
-}
-
-IOReturn itlwm::tsleepHandler(OSObject* owner, void* arg0, void* arg1, void* arg2, void* arg3)
-{
-    itlwm* dev = OSDynamicCast(itlwm, owner);
+    ItlHalService* dev = OSDynamicCast(ItlHalService, owner);
     if (dev == 0)
         return kIOReturnError;
     
@@ -656,8 +625,4 @@ IOReturn itlwm::tsleepHandler(OSObject* owner, void* arg0, void* arg1, void* arg
         else
             return kIOReturnTimeout;
     }
-}
-
-IOReturn itlwm::getMaxPacketSize(UInt32 *maxSize) const {
-    return super::getMaxPacketSize(maxSize);
 }
